refactor: 패키지 구조 변경, 응답코드 검증 부분 공통화 (StatusCode 클래스로 분리)
