+++ src/test/java/nextstep/subway/utils/httpresponse/Header.java
@@ -1,9 +1,11 @@
-package nextstep.subway.fixture;
+package nextstep.subway.utils.httpresponse;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 
-public class CommonFixture {
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class Header {
     public static String uri(ExtractableResponse<Response> response) {
         return response.header("Location");
     }
