+++ src/test/java/nextstep/subway/utils/httpresponse/Response.java
@@ -1,15 +1,14 @@
-package nextstep.subway.utils;
+package nextstep.subway.utils.httpresponse;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import org.springframework.http.MediaType;
 
 import java.util.Map;
 
-public class HttpRequestResponse {
+public class Response {
 
-    public static ExtractableResponse<Response> post(Map<String, String> requestDto, String uri) {
+    public static ExtractableResponse<io.restassured.response.Response> post(Map<String, String> requestDto, String uri) {
         return RestAssured.given().log().all()
                 .body(requestDto)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -19,7 +18,7 @@ public class HttpRequestResponse {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> get(String uri) {
+    public static ExtractableResponse<io.restassured.response.Response> get(String uri) {
         return RestAssured.given().log().all()
                 .when()
                 .get(uri)
@@ -27,7 +26,7 @@ public class HttpRequestResponse {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> put(Map<String, String> requestDto, String uri) {
+    public static ExtractableResponse<io.restassured.response.Response> put(Map<String, String> requestDto, String uri) {
         return RestAssured.given().log().all()
                 .body(requestDto)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -37,7 +36,7 @@ public class HttpRequestResponse {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> delete(String uri) {
+    public static ExtractableResponse<io.restassured.response.Response> delete(String uri) {
         return RestAssured.given().log().all()
                 .when()
                 .delete(uri)
